+++ CHANGELOG.md
@@ -19,3 +19,5 @@
 * Add `inverseOf` for autosaving `Order` relationships
 * Add watcher to customer email changes
 * Increase page size of questions loaded in `Order.construct` to 500
+* Set `answer.option` to default option for `drop_down` and `option_list` questions in `Order.construct`
+* Set `answer.value` to `min` for `spin_button` questions in `Order.construct`